Fixed issues with all the tests

+++ 3DRenderer/forms/tests/test_model.cpp
@@ -15,9 +15,9 @@ TEST(Model, DefaultConstructor)
 TEST(Model, ConstructWithMeshes)
 {
     std::vector<Vertex> verts;
-    verts.push_back(Vertex{Eigen::Vector3f(0.0f, 0.0f, 0.0f), Eigen::Vector3d(1.0, 0.0, 0.0)});
-    verts.push_back(Vertex{Eigen::Vector3f(1.0f, 0.0f, 0.0f), Eigen::Vector3d(0.0, 1.0, 0.0)});
-    verts.push_back(Vertex{Eigen::Vector3f(0.0f, 1.0f, 0.0f), Eigen::Vector3d(0.0, 0.0, 1.0)});
+    verts.push_back(Vertex{Eigen::Vector3f(0.0f, 0.0f, 0.0f), Eigen::Vector3i(1.0, 0.0, 0.0)});
+    verts.push_back(Vertex{Eigen::Vector3f(1.0f, 0.0f, 0.0f), Eigen::Vector3i(0.0, 1.0, 0.0)});
+    verts.push_back(Vertex{Eigen::Vector3f(0.0f, 1.0f, 0.0f), Eigen::Vector3i(0.0, 0.0, 1.0)});
     
     std::vector<Face> faces;
     faces.push_back(Face{0, 1, 2});
@@ -38,7 +38,7 @@ TEST(Model, ConstructWithMeshes)
 TEST(Model, GetMeshesReturnsCorrectData)
 {
     std::vector<Vertex> verts;
-    verts.push_back(Vertex{Eigen::Vector3f(1.0f, 2.0f, 3.0f), Eigen::Vector3d(1.0, 0.0, 0.0)});
+    verts.push_back(Vertex{Eigen::Vector3f(1.0f, 2.0f, 3.0f), Eigen::Vector3i(1.0, 0.0, 0.0)});
     
     std::vector<Face> faces;
     
@@ -115,7 +115,7 @@ TEST(Model, LoadMultipleMeshesEachHasCorrectData)
 TEST(Model, ConstructWithMeshesAndLoadMore)
 {
     std::vector<Vertex> verts;
-    verts.push_back(Vertex{Eigen::Vector3f(0.0f, 0.0f, 0.0f), Eigen::Vector3d(1.0, 0.0, 0.0)});
+    verts.push_back(Vertex{Eigen::Vector3f(0.0f, 0.0f, 0.0f), Eigen::Vector3i(1.0, 0.0, 0.0)});
     
     std::vector<Face> faces;
     
@@ -146,8 +146,8 @@ TEST(Model, EmptyModelGetMeshesReturnsEmptyVector)
 TEST(Model, ConstructedModelPreservesOriginalMeshData)
 {
     std::vector<Vertex> verts;
-    verts.push_back(Vertex{Eigen::Vector3f(1.0f, 2.0f, 3.0f), Eigen::Vector3d(1.0, 0.0, 0.0)});
-    verts.push_back(Vertex{Eigen::Vector3f(4.0f, 5.0f, 6.0f), Eigen::Vector3d(0.0, 1.0, 0.0)});
+    verts.push_back(Vertex{Eigen::Vector3f(1.0f, 2.0f, 3.0f), Eigen::Vector3i(1.0, 0.0, 0.0)});
+    verts.push_back(Vertex{Eigen::Vector3f(4.0f, 5.0f, 6.0f), Eigen::Vector3i(0.0, 1.0, 0.0)});
     
     std::vector<Face> faces;
     faces.push_back(Face{0, 1, 0});
